*move ScaledDotProductAttentionLayer to Layers/Normalize.
