+++ prj/vs2022/Synet.vcxproj.filters
@@ -26,8 +26,8 @@
     <ClCompile Include="..\..\src\Synet\Layers\NonZeroLayer.cpp">
       <Filter>Layers</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\Synet\Layers\ScaledDotProductAttentionLayer.cpp">
-      <Filter>Layers</Filter>
+    <ClCompile Include="..\..\src\Synet\Layers\Normalize\ScaledDotProductAttentionLayer.cpp">
+      <Filter>Layers\Normalize</Filter>
     </ClCompile>
     <ClCompile Include="..\..\src\Synet\Layers\Legacy\BroadcastLayer.cpp">
       <Filter>Layers\Legacy</Filter>
@@ -431,8 +431,8 @@
     <ClInclude Include="..\..\src\Synet\Layers\NonZeroLayer.h">
       <Filter>Layers</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\src\Synet\Layers\ScaledDotProductAttentionLayer.h">
-      <Filter>Layers</Filter>
+    <ClInclude Include="..\..\src\Synet\Layers\Normalize\ScaledDotProductAttentionLayer.h">
+      <Filter>Layers\Normalize</Filter>
     </ClInclude>
     <ClInclude Include="..\..\src\Synet\Decoders\Rtdetr.h">
       <Filter>Decoders</Filter>
